guard RefreshPeer():sync() not to insert the same record twice.
The problem is sync() inserts duplicated records, but when we try to
insert into peer_* tables it fails for the second one. That creates
strange problems like federated objects appear as local objects.

We still need to find why sync is called with duplicated records but
this patch should prevent the db problems.